client-requirements.txt
flwr>=1.10,<1.11
fastapi>=0.115,<1
numpy>=1.26,<3
pandas>=2.2,<3
python-dotenv>=1.0,<2
scikit-learn>=1.4,<2
torch>=2.2,<3
opacus>=1.5,<2
uvicorn>=0.30,<1
